Update files paths
