--- link.py
@@ -1,8 +1,0 @@
-url = ["https://merlin.pl/catalog/ksiazki-m10349074/kryminal-i-sensacja-c495/",
-       "https://merlin.pl/catalog/ksiazki-m10349074/romanse-c1774/",
-       "https://merlin.pl/catalog/ksiazki-m10349074/historia-c100144/",
-       "https://merlin.pl/catalog/ksiazki-m10349074/fantastyka-c467/",
-       "https://merlin.pl/catalog/ksiazki-m10349074/literatura-faktu-c100173/",
-       "https://merlin.pl/catalog/ksiazki-m10349074/poezja-i-dramat-c5869/",
-       "https://merlin.pl/catalog/ksiazki-m10349074/sztuka-c274/",
-       "https://merlin.pl/catalog/ksiazki-m10349074/biznes-ekonomia-c100118/"]
