+++ scripts/scraper/book.py
@@ -1,0 +1,12 @@
+class book:
+    def __init__(self, title, author, publisher, category, description, informations, price, currentPrice, status, img):
+        self.title = title
+        self.author = author
+        self.publisher = publisher
+        self.category = category
+        self.description = description
+        self.informations = informations
+        self.price = price
+        self.currentPrice = currentPrice
+        self.status = status
+        self.img = img
